Moves files.
