--- Packages/OpenVanilla/Sources/OpenVanillaImp/include/OpenVanillaImp.h
@@ -1,9 +1,0 @@
-#include "OVCandidateListImpl.h"
-#include "OVCandidateServiceImpl.h"
-#include "OVConcreteKeyImpl.h"
-#include "OVLoaderServiceImpl.h"
-#include "OVOneDimensionalCandidatePanelImpl.h"
-#include "OVPlistBackedKeyValueMapImpl.h"
-#include "OVTextBufferCombinator.h"
-#include "OVTextBufferImpl.h"
-#include "OVConstants.h"
