+++ Source/Mac/OpenVanilla-Bridging-Header.h
@@ -2,6 +2,16 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-//#import "BasePreferencesViewController.h"
 
 #import "OVInputSourceHelper.h"
+#import "OVLoaderServiceImpl.h"
+#import "OVModuleManager.h"
+#import "OVCandidateListImpl.h"
+#import "OVCandidateServiceImpl.h"
+#import "OVConcreteKeyImpl.h"
+#import "OVLoaderServiceImpl.h"
+#import "OVOneDimensionalCandidatePanelImpl.h"
+#import "OVPlistBackedKeyValueMapImpl.h"
+#import "OVTextBufferCombinator.h"
+#import "OVTextBufferImpl.h"
+#import "OVConstants.h"
