+++ Source/Mac/OpenVanillaImpl/OVCandidateListImpl.h
