+++ Source/Mac/OpenVanillaImpl/OVCandidateServiceImpl.h
