+++ Source/Mac/OpenVanillaImpl/OVOneDimensionalCandidatePanelImpl.h
@@ -27,7 +27,8 @@
 
 #import "OVCandidateService.h"
 #import "OVCandidateListImpl.h"
-#import "VTCandidateController.h"
+//#import "VTCandidateController.h"
+@import Voltaire;
 
 extern NSString *const OVOneDimensionalCandidatePanelImplDidSelectCandidateNotification;
 extern NSString *const OVOneDimensionalCandidatePanelImplSelectedCandidateIndexKey;
