+++ Source/Mac/OpenVanillaImpl/OVTextBufferCombinator.h
