+++ Source/Mac/OpenVanillaImpl/OVTextBufferCombinator.mm
