+++ Source/Mac/OpenVanillaImpl/OVTextBufferImpl.cpp
