+++ Source/Mac/OpenVanillaImpl/OVTextBufferImpl.h
